rename pkg/cstest -> pkg/hubtest, keep cstest for generic helper functions
this also avoids circular imports in test files
