+++ pkg/hubtest/hubtest_item.go
@@ -1,4 +1,4 @@
-package cstest
+package hubtest
 
 import (
 	"fmt"
